Add --uniform-space option to choose uniform space
This is useful to e.g. run viscm with --uniform-space=CIELab and check
that parula is indeed uniform in L*a*b* space.

It is also useful because it turns out that I had a stupid bug in the
colorspace conversion code that meant we were not *quite* using the
correct uniform space for our calculations. See:

  https://github.com/njsmith/colorspacious/commit/9a4d871bdc2f541239b2a307da00fff627088fa7

So now you can pass --uniform-space=buggy-CAM02-UCS to reproduce the old
results. Fortunately if we use the fixed code to evaluate the old
colormaps then they are fine -- they aren't *quite* perceptually uniform
anymore, but the deviations are not enough to care about. But if you
want to edit the old colormaps or to reproduce them from their spline
representation, then you do need this command-line option, because
fixing the bug caused the sRGB solid in CAM02-UCS space to be slightly
rescaled, so suddenly our existing CAM02-UCS coordinates are mostly
out-of-gamut.