FIX: respect 'bytes' kwarg
In mpl 2.0 the value of the 'bytes' kwarg to `ColorMap.__call__`
while making an image.

The viscm transform code was assuming that it get a float result from
the call to base_cmap() and the external code was expected back uint8
values.

The 'bytes' kwarg is now caught and applied on the way out if needed.