Fix failing viscm editor test
The old behavior was an implicit conversion to `int`. This warning was
being thrown in Python 3.9:

```
DeprecationWarning: an integer is required (got type float).  Implicit conversion to integers using __int__ is deprecated, and may be removed in a future version of Python.
```

And in Python 3.10, this turned in to an error, as seen in
<https://github.com/matplotlib/viscm/issues/71>:

```
TypeError: setValue(self, a0: int): argument 1 has unexpected type 'float'
```

The new version of this code simply makes the old behavior (conversion
to int) explicit. I'm unsure whether this was the original intent.